feat: 댓글 조회, 작성, 수정, 삭제 기능 구현

+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/dto/CommentResponse.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CommentResponse(
-    val commentId: Long,
-    val memberId: Long,
+    val commentId: Int,
+    val memberId: Int,
     val nickname: String,
     val content: String,
     val createdAt: String,
