+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/model/CommentModel.kt
@@ -1,8 +1,8 @@
 package com.yeobaek.data.model
 
 data class CommentModel(
-    val commentId: Long,
-    val memberId: Long,
+    val commentId: Int,
+    val memberId: Int,
     val nickname: String,
     val content: String,
     val createdAt: String,
