+++ android/shared/src/commonMain/kotlin/com/yeobaek/feature/reader/ReaderScreen.kt
@@ -48,9 +48,9 @@ fun ReaderScreen(
     onCommentSheetDismiss: () -> Unit,
     onCommentInputChange: (String) -> Unit,
     onCommentSubmit: () -> Unit,
-    onCommentEdit: (Long) -> Unit,
+    onCommentEdit: (Int) -> Unit,
     onCommentEditCancel: () -> Unit,
-    onCommentDelete: (Long) -> Unit,
+    onCommentDelete: (Int) -> Unit,
     onCommentDeleteCancel: () -> Unit,
     onCommentDeleteConfirm: () -> Unit,
     onLoadPrevious: () -> Boolean,
@@ -409,7 +409,7 @@ private fun ReaderScreenPreview() {
 }
 
 internal data class PassageAnchor(
-    val passageId: Long,
+    val passageId: Int,
     val scrollOffset: Int,
 )
